registration-system/server/requirements.txt
flask==2.0.1
pymongo==4.0.1
python-dotenv==0.19.0
requests==2.26.0
azure-identity==1.7.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.0.1
pyjwt==2.8.0 